requirements.txt
cloudinary
Django
django-admin-thumbnails
django-cors-headers
djangorestframework
gunicorn
numpy
Pillow
psycopg2-binary
python-decouple
tflite-runtime
whitenoise
